+++ lib/millis/src/millis.h
@@ -1,7 +1,9 @@
 #ifndef LIB_MILLIS_H
 #define LIB_MILLIS_H
 
-unsigned long millis(void);
+#include "stdint.h"
+
+uint16_t millis(void);
 void init_millis(void);
 
 #endif // LIB_MILLIS_H
